Move DateTimeExtensions to own file
