+++ GitCommands/GitCommands.csproj
@@ -72,6 +72,7 @@
     <Compile Include="CommitTemplateItem.cs" />
     <Compile Include="CommitTemplateManager.cs" />
     <Compile Include="Config\SettingKeyString.cs" />
+    <Compile Include="DateTimeExtensions.cs" />
     <Compile Include="DateTimeUtils.cs" />
     <Compile Include="DeconstructionExtensions.cs" />
     <Compile Include="EnvironmentAbstraction.cs" />
